ACM/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
requests>=2.25.0
matplotlib>=3.3.0
gpxpy>=1.4.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.0.0
shapely>=2.0.0
